UF-16037: Added process model
