+++ src/main/java/se/sundsvall/paratransit/Constants.java
@@ -1,0 +1,16 @@
+package se.sundsvall.paratransit;
+
+import generated.se.sundsvall.camunda.VariableValueDto;
+import org.camunda.bpm.engine.variable.type.ValueType;
+
+public class Constants {
+
+	private Constants() {}
+
+	public static final String PROCESS_KEY = "process-paratransit"; // Must match ID of process defined in bpmn schema
+	public static final String TENANTID_TEMPLATE = "PARATRANSIT"; // Namespace where process is deployed, a.k.a tenant (must match setting in application.yaml)
+
+	public static final String UPDATE_AVAILABLE = "updateAvailable";
+	public static final VariableValueDto TRUE = new VariableValueDto().type(ValueType.BOOLEAN.getName()).value(true);
+	public static final VariableValueDto FALSE = new VariableValueDto().type(ValueType.BOOLEAN.getName()).value(false);
+}
