+++ src/main/java/se/sundsvall/paratransit/integration/camunda/configuration/CamundaProperties.java
@@ -1,0 +1,6 @@
+package se.sundsvall.paratransit.integration.camunda.configuration;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties("integration.camunda")
+public record CamundaProperties(int connectTimeout, int readTimeout) {}
