+++ src/main/java/se/sundsvall/paratransit/integration/camunda/deployment/DeploymentException.java
@@ -1,0 +1,9 @@
+package se.sundsvall.paratransit.integration.camunda.deployment;
+
+public class DeploymentException extends RuntimeException {
+	private static final long serialVersionUID = -1616889424590166876L;
+
+	public DeploymentException(Exception e) {
+		super(e);
+	}
+}
