+++ src/main/java/se/sundsvall/paratransit/service/ProcessService.java
@@ -1,0 +1,39 @@
+package se.sundsvall.paratransit.service;
+
+import static org.zalando.problem.Status.NOT_FOUND;
+import static se.sundsvall.paratransit.Constants.PROCESS_KEY;
+import static se.sundsvall.paratransit.Constants.TENANTID_TEMPLATE;
+import static se.sundsvall.paratransit.Constants.TRUE;
+import static se.sundsvall.paratransit.Constants.UPDATE_AVAILABLE;
+
+import generated.se.sundsvall.camunda.StartProcessInstanceDto;
+import org.springframework.stereotype.Service;
+import org.zalando.problem.Problem;
+import se.sundsvall.paratransit.integration.camunda.CamundaClient;
+
+@Service
+public class ProcessService {
+
+	private final CamundaClient camundaClient;
+
+	public ProcessService(CamundaClient camundaClient) {
+		this.camundaClient = camundaClient;
+	}
+
+	public String startProcess(String businessKey) {
+		return camundaClient.startProcessWithTenant(PROCESS_KEY, TENANTID_TEMPLATE, new StartProcessInstanceDto().businessKey(businessKey)).getId();
+	}
+
+	public void updateProcess(String processInstanceId) {
+
+		verifyExistingProcessInstance(processInstanceId);
+
+		camundaClient.setProcessInstanceVariable(processInstanceId, UPDATE_AVAILABLE, TRUE);
+	}
+
+	private void verifyExistingProcessInstance(String processInstanceId) {
+		if (camundaClient.getProcessInstance(processInstanceId).isEmpty()) {
+			throw Problem.valueOf(NOT_FOUND, "Process instance with ID '%s' does not exist!".formatted(processInstanceId));
+		}
+	}
+}
