--- src/main/java/se/sundsvall/workflow/integration/camunda/configuration/CamundaProperties.java
@@ -1,6 +1,0 @@
-package se.sundsvall.workflow.integration.camunda.configuration;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties("integration.camunda")
-public record CamundaProperties(int connectTimeout, int readTimeout) {}
