--- src/main/java/se/sundsvall/workflow/integration/camunda/deployment/DeploymentException.java
@@ -1,9 +1,0 @@
-package se.sundsvall.workflow.integration.camunda.deployment;
-
-public class DeploymentException extends RuntimeException {
-	private static final long serialVersionUID = -1616889424590166876L;
-
-	public DeploymentException(Exception e) {
-		super(e);
-	}
-}
