+++ src/main/resources/application-it.yaml
@@ -1,6 +1,14 @@
+#========================================
+# Integration test specific settings
+#========================================
+#----------------------------------------
+# Spring settings
+#----------------------------------------
 spring:
   application:
-    name: template-camunda-it
+    name: pw-paratransit-it
+  main:
+    banner-mode: "off"
 #----------------------------------------
 # Open API
 #----------------------------------------
@@ -9,28 +17,11 @@ openapi:
   title: ${spring.application.name}
   version: '@project.version@'
 
-#----------------------------------------
-# Camunda
-#----------------------------------------
-
-camunda:
-  bpm:
-    client:
-      base-url: http://replaced_by_test:8080
-    deployment:
-      processes:
-        - name: Template process
-          tenant: TEMPLATE_NAMESPACE
-          bpmnResourcePattern: classpath*:processmodels/*.bpmn
-          dmnResourcePattern: classpath*:processmodels/*.dmn
-          formResourcePattern: classpath*:processmodels/*.form
   
 #----------------------------------------
 # Camunda integration settings
 #----------------------------------------
 # Rest client
-integration:
+config:
   camunda:
-    connectTimeout: 5
-    readTimeout: 20
-    url: http://replaced_by_test:8080
+    base-url: http://localhost:${wiremock.server.port}/api-camunda/engine-rest
