+++ src/main/resources/application-junit.yaml
@@ -2,7 +2,13 @@
 # Unit test specific settings
 #========================================
 #----------------------------------------
-# Camunda
+# Spring settings
+#----------------------------------------
+spring:
+  main:
+    banner-mode: "off"
+#----------------------------------------
+# Camunda settings
 #----------------------------------------
 camunda:
   bpm:
