+++ src/main/resources/application.yaml
@@ -1,8 +1,6 @@
-server:
-  port: 8081
 spring:
   application:
-    name: template-camunda
+    name: pw-paratransit
 #----------------------------------------
 # Open API
 #----------------------------------------
@@ -17,11 +15,15 @@ openapi:
 camunda:
   bpm:
     client:
-      base-url: http://localhost:8080/engine-rest
+      base-url: ${config.camunda.base-url}
+      backoff:
+        initTime: 500
+        factor: 2
+        maxTime: 15000
     deployment:
       processes:
-        - name: Template process
-          tenant: TEMPLATE_NAMESPACE
+        - name: Paratransit process
+          tenant: PARATRANSIT
           bpmnResourcePattern: classpath*:processmodels/*.bpmn
           dmnResourcePattern: classpath*:processmodels/*.dmn
           formResourcePattern: classpath*:processmodels/*.form
@@ -39,7 +41,7 @@ integration:
   camunda:
     connectTimeout: 5
     readTimeout: 20
-    url: http://localhost:8080/engine-rest
+    url: ${config.camunda.base-url}
 
 #----------------------------------------
 # Circuit breakers
