+++ src/test/java/se/sundsvall/paratransit/integration/camunda/configuration/CamundaConfigurationTest.java
@@ -1,0 +1,57 @@
+package se.sundsvall.paratransit.integration.camunda.configuration;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static se.sundsvall.paratransit.integration.camunda.configuration.CamundaConfiguration.CLIENT_ID;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Mock;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
+import org.mockito.Spy;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.cloud.openfeign.FeignBuilderCustomizer;
+import se.sundsvall.dept44.configuration.feign.FeignMultiCustomizer;
+import se.sundsvall.dept44.configuration.feign.decoder.ProblemErrorDecoder;
+
+@ExtendWith(MockitoExtension.class)
+class CamundaConfigurationTest {
+
+	@Spy
+	private FeignMultiCustomizer feignMultiCustomizerSpy;
+
+	@Mock
+	private FeignBuilderCustomizer feignBuilderCustomizerMock;
+
+	@Mock
+	private CamundaProperties propertiesMock;
+
+	@Test
+	void testFeignBuilderCustomizer() {
+		final var configuration = new CamundaConfiguration();
+
+		when(propertiesMock.connectTimeout()).thenReturn(1);
+		when(propertiesMock.readTimeout()).thenReturn(2);
+		when(feignMultiCustomizerSpy.composeCustomizersToOne()).thenReturn(feignBuilderCustomizerMock);
+
+		try (MockedStatic<FeignMultiCustomizer> feignMultiCustomizerMock = Mockito.mockStatic(FeignMultiCustomizer.class)) {
+			feignMultiCustomizerMock.when(FeignMultiCustomizer::create).thenReturn(feignMultiCustomizerSpy);
+
+			var customizer = configuration.feignBuilderCustomizer(propertiesMock);
+
+			ArgumentCaptor<ProblemErrorDecoder> errorDecoderCaptor = ArgumentCaptor.forClass(ProblemErrorDecoder.class);
+
+			verify(feignMultiCustomizerSpy).withErrorDecoder(errorDecoderCaptor.capture());
+			verify(propertiesMock).connectTimeout();
+			verify(propertiesMock).readTimeout();
+			verify(feignMultiCustomizerSpy).withRequestTimeoutsInSeconds(1, 2);
+			verify(feignMultiCustomizerSpy).composeCustomizersToOne();
+
+			assertThat(errorDecoderCaptor.getValue()).hasFieldOrPropertyWithValue("integrationName", CLIENT_ID);
+			assertThat(customizer).isSameAs(feignBuilderCustomizerMock);
+		}
+	}
+}
