+++ src/test/java/se/sundsvall/paratransit/integration/camunda/configuration/CamundaPropertiesTest.java
@@ -1,0 +1,23 @@
+package se.sundsvall.paratransit.integration.camunda.configuration;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import se.sundsvall.paratransit.Application;
+
+@SpringBootTest(classes = Application.class)
+@ActiveProfiles("junit")
+class CamundaPropertiesTest {
+
+	@Autowired
+	private CamundaProperties properties;
+
+	@Test
+	void testProperties() {
+		assertThat(properties.connectTimeout()).isEqualTo(5);
+		assertThat(properties.readTimeout()).isEqualTo(20);
+	}
+}
