+++ src/test/java/se/sundsvall/paratransit/integration/camunda/deployment/DeploymentExceptionTest.java
@@ -1,0 +1,13 @@
+package se.sundsvall.paratransit.integration.camunda.deployment;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.Test;
+
+class DeploymentExceptionTest {
+
+	@Test
+	void testInheritance() {
+		assertThat(DeploymentException.class).hasSuperclass(RuntimeException.class);
+	}
+}
