--- src/test/java/se/sundsvall/workflow/api/model/StartProcessResponseTest.java
@@ -1,40 +1,0 @@
-package se.sundsvall.workflow.api.model;
-
-import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanConstructor;
-import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanEquals;
-import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanHashCode;
-import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanToString;
-import static com.google.code.beanmatchers.BeanMatchers.hasValidGettersAndSetters;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.CoreMatchers.allOf;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-import org.junit.jupiter.api.Test;
-
-class StartProcessResponseTest {
-
-	@Test
-	void testBean() {
-		assertThat(StartProcessResponse.class, allOf(
-			hasValidBeanConstructor(),
-			hasValidGettersAndSetters(),
-			hasValidBeanHashCode(),
-			hasValidBeanEquals(),
-			hasValidBeanToString()));
-	}
-
-	@Test
-	void testBuilderMethods() {
-		final var processInstanceId = "processInstanceId";
-
-		final var bean = new StartProcessResponse(processInstanceId);
-
-		assertThat(bean).isNotNull().hasNoNullFieldsOrProperties();
-		assertThat(bean.getProcessId()).isEqualTo(processInstanceId);
-	}
-
-	@Test
-	void testNoDirtOnCreatedBean() {
-		assertThat(new StartProcessResponse()).hasAllNullFieldsOrProperties();
-	}
-}
