--- src/test/java/se/sundsvall/workflow/integration/camunda/configuration/CamundaPropertiesTest.java
@@ -1,23 +1,0 @@
-package se.sundsvall.workflow.integration.camunda.configuration;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import se.sundsvall.workflow.Application;
-
-@SpringBootTest(classes = Application.class)
-@ActiveProfiles("junit")
-class CamundaPropertiesTest {
-
-	@Autowired
-	private CamundaProperties properties;
-
-	@Test
-	void testProperties() {
-		assertThat(properties.connectTimeout()).isEqualTo(5);
-		assertThat(properties.readTimeout()).isEqualTo(20);
-	}
-}
